requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
Pillow==10.0.1
djangorestframework~=3.14.0
django-filter~=23.3
psycopg2-binary==2.9.7
pytz==2023.3.post1

stripe~=7.0.0
pytest~=7.4.2
celery~=5.3.4